requirements.txt
flask>=2.3.0
torch>=2.0.0
numpy>=1.24.0
python-chess>=3.0.0

